Restructuration

+++ sources/Stockage/managers/StockageBDD.cs
@@ -2,6 +2,8 @@
 using SQLite;
 using System.IO;
 using Models;
+using System.Collections.ObjectModel;
+using System.Collections.Generic;
 
 namespace Managers
 {
@@ -9,9 +11,12 @@ namespace Managers
     {
         private const string DatabaseFile = "db.sqlite";
         public static SQLiteConnection Database { get; private set; }
-        
-        public static void Init()
+
+        public static SQLiteConnection Init()
         {
+            if (Database != null)
+                return Database;
+
             File.Delete(DatabaseFile);
 
             Database = new SQLiteConnection(DatabaseFile);
@@ -21,6 +26,8 @@ namespace Managers
 
             Database.CreateTable<Utilisateur>();
             Database.CreateTable<Film>();
+
+            return Database;
         }
 
         public static int Insert<T>(T u)
@@ -33,6 +40,13 @@ namespace Managers
             var query = Database.Table<Utilisateur>().Where(u => u.Pseudo.Equals(pseudo) && u.Password.Equals(mdp));
 
             return query.Count() == 1;
+        }
+         
+        public static ObservableCollection<Film> GetFilms()
+        {
+            var list = Database.Query<Film>("SELECT * FROM Film ORDER BY titre");
+
+            return new ObservableCollection<Film>(list);
         }
     }
